improve error handling

+++ weather-todo-app/src/features/list-todo/delete-todo.js
@@ -1,11 +1,20 @@
 import { deleteOne } from "./list-todo-slice.js";
+import { showError } from "../error/error-slice.js";
 
 export default function deleteTodo(uuid, store) {
-  fetch(`/api/todo/delete/${uuid}`, {
-    method: "DELETE",
-  })
-    .then((res) => res.json())
-    .then((result) => {
-      store.dispatch(deleteOne(result.data));
-    });
+  try {
+    fetch(`/api/todo/delete/${uuid}`, {
+      method: "DELETE",
+    })
+      .then((res) => res.json())
+      .then((result) => {
+        if (result.code === 200) {
+          store.dispatch(deleteOne(result.data));
+        } else {
+          store.dispatch(showError());
+        }
+      });
+  } catch (e) {
+    throw new Error(e, "Server error");
+  }
 }
